Model_Deployment/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
fastapi[standard]
python-multipart
ultralytics
Pillow